Moving process struct to library
- Moving performance struct to library.
