+++ performance-queue/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "performance-queue"
+version = "0.1.0"
+edition = "2021"
+
+authours = ["rhasler1 <rhasler@luc.edu>"]
+
+[dependencies]
