+++ process-list/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "process-list"
+version = "0.1.0"
+edition = "2021"
+
+authours = ["rhasler1 <rhasler@luc.edu>"]
+
+[dependencies]
+itertools = "0.10.0"
