+++ src/components/process.rs
@@ -5,11 +5,11 @@ use ratatui::{
     prelude::*,
     widgets::{block::*, *},
 };
+use process_list::{ProcessList, ProcessListItems, ProcessListItem};
+
+use super::{common_nav, common_sort};
 use super::{filter::FilterComponent, Component, DrawableComponent, EventState};
 use super::utils::vertical_scroll::VerticalScroll;
-use crate::process_structs::{common_nav, common_sort, process_list_items::ProcessListItems};
-use crate::process_structs::process_list_item::ProcessListItem;
-use crate::process_structs::process_list::ProcessList;
 use crate::config::KeyConfig;
 
 // The ProcessComponent can be navigated to focus on
@@ -19,8 +19,6 @@ use crate::config::KeyConfig;
 pub enum Focus {
     Filter,
     List,
-    Terminate,
-    // Add Terminate variant; This way we can match focus when drawing the ProcessComponent and draw TerminateComponent.
 }
 
 pub struct ProcessComponent {
@@ -66,13 +64,124 @@ impl ProcessComponent {
     pub fn selected_pid(&self) -> Option<u32> {
         if matches!(self.focus, Focus::List) {
             if let Some(list) = self.filtered_list.as_ref() {
-                return list.get_selected_pid()
+                return list.selected_pid()
             }
             else {
-                return self.list.get_selected_pid()
+                return self.list.selected_pid()
             }
         }
         None
+    }
+
+    fn draw_process_list(&mut self, f: &mut Frame, area: Rect, _focused: bool) -> io::Result<()> {
+        // Setting the list height to the height of the vertical chunk for the process list; We are subtracting
+        // two from the height to account for the area that will be taken up by the border around the list.
+        let visual_list_height = (area.height.saturating_sub(2)) as usize;
+
+        // Getting the list to display; If there is some filtered list display it, else display the unfiltered list.
+        let list = if let Some(list) = self.filtered_list.as_ref() {
+            list
+        }
+        else {
+            &self.list
+        };
+
+        // Updating the scroll struct which calculates the position at the top of the displayed list.
+        list.selection().map_or_else(
+            { ||
+                self.scroll.reset()
+            }, |selection| {
+                self.scroll.update(
+                    selection,list.len(), visual_list_height
+                );
+            },
+        );
+
+        // Getting the boolean list.follow(); The follow_flag is used to differentiate between a selected item being followed(underlined) and not.
+        let follow_flag = list.follow();
+
+        // Different styles used to visually differentiate between components and focus.
+        let header_style = Style::default().fg(Color::Black).bg(Color::Gray);
+        let select_style = Style::default().bg(Color::Blue).add_modifier(Modifier::BOLD);
+        let select_follow_style = Style::default().bg(Color::Blue).add_modifier(Modifier::BOLD).add_modifier(Modifier::UNDERLINED);
+        let default_style = Style::default().fg(Color::White);
+        let out_of_focus_style = Style::default().fg(Color::DarkGray);
+
+        // Setting the header.
+        let header = ["Pid", "Name", "Cpu Usage (%)", "Memory Usage (Bytes)"]
+            .into_iter()
+            .map(Cell::from)
+            .collect::<Row>()
+            .style(
+                if matches!(self.focus, Focus::List) {
+                    header_style
+                }
+                else {
+                    out_of_focus_style
+                }
+            )
+            .height(1);
+
+        // Setting the rows to display; The iterate function iterates starting from the value self.scroll.get_top() and a list_height number of times.
+        // We don't iterate over the entire list everytime we draw the list, instead we only iterate over the portion that is being displayed.
+        // See process_structs::list_items_iter::next for the implementation.
+        let rows = list
+            .iterate(self.scroll.get_top(), visual_list_height)
+            .map(|(item, selected)| {
+                let style =
+                    if matches!(self.focus, Focus::List) && selected && follow_flag {
+                        select_follow_style
+                    }
+                    else if matches!(self.focus, Focus::List) && selected && !follow_flag {
+                        select_style
+                    }
+                    else if matches!(self.focus, Focus::List) {
+                        default_style
+                    }
+                    else {
+                        out_of_focus_style
+                    };
+
+                let cells = vec![
+                    Cell::from(item.pid().to_string()),
+                    Cell::from(item.name().to_string()),
+                    Cell::from(item.cpu_usage().to_string()),
+                    Cell::from(item.memory_usage().to_string()),
+                ];
+
+                Row::new(cells).style(style)
+            })
+            .collect::<Vec<_>>();
+
+        // Setting the width constraints.
+        let widths =
+            vec![
+                Constraint::Length(10),
+                Constraint::Length(50),
+                Constraint::Length(20),
+                Constraint::Length(20),
+            ];
+
+        // Setting block information.
+        let block_title: &str = "Process List";
+        let block_style =
+            if matches!(self.focus, Focus::List) {
+                Style::default().fg(Color::White)
+            }
+            else {
+                Style::default().fg(Color::DarkGray)
+            };
+
+        // Setting the table.
+        let table = Table::new(rows, widths)
+            .header(header)
+            .block(Block::default().borders(Borders::ALL).title(block_title))
+            .style(block_style);
+
+        // Render.
+        f.render_widget(table, area);
+
+        Ok(())
     }
 }
 
@@ -208,112 +317,8 @@ impl DrawableComponent for ProcessComponent {
         // Drawing the filter component.
         self.filter.draw(f, horizontal_chunks[1], matches!(self.focus, Focus::Filter))?;
 
-        // Setting the list height to the height of the vertical chunk for the process list; We are subtracting
-        // two from the height to account for the area that will be taken up by the border around the list.
-        let visual_list_height = (vertical_chunks[1].height.saturating_sub(2)) as usize;
-
-        // Getting the list to display; If there is some filtered list display it, else display the unfiltered list.
-        let list = if let Some(list) = self.filtered_list.as_ref() {
-            list
-        }
-        else {
-            &self.list
-        };
-
-        // Updating the scroll struct which calculates the position at the top of the displayed list.
-        list.selection().map_or_else(
-            { ||
-                self.scroll.reset()
-            }, |selection| {
-                self.scroll.update(
-                    selection,list.list_len(), visual_list_height
-                );
-            },
-        );
-
-        // Getting the boolean list.follow(); The follow_flag is used to differentiate between a selected item being followed(underlined) and not.
-        let follow_flag = list.follow();
-
-        // Different styles used to visually differentiate between components and focus.
-        let header_style = Style::default().fg(Color::Black).bg(Color::Gray);
-        let select_style = Style::default().bg(Color::Blue).add_modifier(Modifier::BOLD);
-        let select_follow_style = Style::default().bg(Color::Blue).add_modifier(Modifier::BOLD).add_modifier(Modifier::UNDERLINED);
-        let default_style = Style::default().fg(Color::White);
-        let out_of_focus_style = Style::default().fg(Color::DarkGray);
-
-        // Setting the header.
-        let header = ["Pid", "Name", "Cpu Usage (%)", "Memory Usage (Bytes)"]
-            .into_iter()
-            .map(Cell::from)
-            .collect::<Row>()
-            .style(
-                if matches!(self.focus, Focus::List) {
-                    header_style
-                }
-                else {
-                    out_of_focus_style
-                }
-            )
-            .height(1);
-
-        // Setting the rows to display; The iterate function iterates starting from the value self.scroll.get_top() and a list_height number of times.
-        // We don't iterate over the entire list everytime we draw the list, instead we only iterate over the portion that is being displayed.
-        // See process_structs::list_items_iter::next for the implementation.
-        let rows = list
-            .iterate(self.scroll.get_top(), visual_list_height)
-            .map(|(item, selected)| {
-                let style =
-                    if matches!(self.focus, Focus::List) && selected && follow_flag {
-                        select_follow_style
-                    }
-                    else if matches!(self.focus, Focus::List) && selected && !follow_flag {
-                        select_style
-                    }
-                    else if matches!(self.focus, Focus::List) {
-                        default_style
-                    }
-                    else {
-                        out_of_focus_style
-                    };
-
-                let cells = vec![
-                    Cell::from(item.pid().to_string()),
-                    Cell::from(item.name().to_string()),
-                    Cell::from(item.cpu_usage().to_string()),
-                    Cell::from(item.memory_usage().to_string()),
-                ];
-
-                Row::new(cells).style(style)
-            })
-            .collect::<Vec<_>>();
-
-        // Setting the width constraints.
-        let widths =
-            vec![
-                Constraint::Length(10),
-                Constraint::Length(50),
-                Constraint::Length(20),
-                Constraint::Length(20),
-            ];
-
-        // Setting block information.
-        let block_title: &str = "Process List";
-        let block_style =
-            if matches!(self.focus, Focus::List) {
-                Style::default().fg(Color::White)
-            }
-            else {
-                Style::default().fg(Color::DarkGray)
-            };
-
-        // Setting the table.
-        let table = Table::new(rows, widths)
-            .header(header)
-            .block(Block::default().borders(Borders::ALL).title(block_title))
-            .style(block_style);
-
-        // Render.
-        f.render_widget(table, vertical_chunks[1]);
+        // Draw process list.
+        self.draw_process_list(f, vertical_chunks[1], matches!(self.focus, Focus::List))?;
 
         // Draw scrollbar.
         self.scroll.draw(f, vertical_chunks[1], false)?;
@@ -324,5 +329,5 @@ impl DrawableComponent for ProcessComponent {
 
 #[cfg(test)]
 mod test {
-    // TODO: write tests, then write TerminateComponent.
+    // TODO: write tests
 }
