+++ src/main.rs
@@ -18,8 +18,6 @@ pub mod app;
 pub mod config;
 pub mod components;
 pub mod events;
-pub mod process_structs;
-pub mod performance_structs;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
