--- src/performance_structs/mod.rs
@@ -1,3 +1,0 @@
-pub mod perf_item;
-pub mod perf_items;
-pub mod cpu_perf_info;
